+++ src/core/index.ts
@@ -11,7 +11,6 @@ import { SUPPORTED_CHAIN_ID } from "../common/chain";
 import { getGasPriceForChain } from "../common/gas-price";
 import { invariant } from "../common/invariant";
 import { AppModule, ModuleMetadataNoType } from "../modules/app";
-import { CollectionModule } from "../modules/collection";
 import {BundleModule} from "../modules/bundle";
 import { DatastoreModule } from "../modules/datastore";
 import { DropModule } from "../modules/drop";
@@ -82,7 +81,6 @@ export interface ISDKOptions {
  */
 export type AnyContract =
   | typeof AppModule
-  | typeof CollectionModule
   | typeof BundleModule
   | typeof NFTModule
   | typeof CurrencyModule
@@ -338,8 +336,8 @@ export class ThirdwebSDK {
    * @param address - The contract address of the given Bundle module.
    * @returns The Bundle Module.
    */
-  public getCollectionModule(address: string): CollectionModule {
-    return this.getOrCreateModule(address, CollectionModule);
+  public getCollectionModule(address: string): BundleModule {
+    return this.getOrCreateModule(address, BundleModule);
   }
   public getBundleModule(address: string): BundleModule {
     return this.getOrCreateModule(address, BundleModule);
