+++ test/collection.test.ts
@@ -1,13 +1,14 @@
-import { CollectionModule, ThirdwebSDK } from "../src/index";
+import { BundleModule, ThirdwebSDK } from "../src/index";
 import * as chai from "chai";
 
 global.fetch = require("node-fetch");
 
 const RPC_URL = "https://matic-mumbai.chainstacklabs.com";
 
-describe("Collection Module", async () => {
+describe("Bundle Module (aka Collection Module)", async () => {
   let sdk: ThirdwebSDK;
-  let collectionModule: CollectionModule;
+  let collectionModule: BundleModule;
+  let BundleModule: BundleModule;
 
   beforeEach(async () => {
     sdk = new ThirdwebSDK(RPC_URL);
@@ -17,6 +18,11 @@ describe("Collection Module", async () => {
      * It contains some test data with burned tokens and some tokens owned by
      * the test address starting with 0xE79
      */
+
+    BundleModule = sdk.getBundleModule(
+      "0x6Da734b14e4CE604f1e18efb7E7f7ef022e96616",
+    )
+    /*Testing backwards compatibility*/
     collectionModule = sdk.getCollectionModule(
       "0x6Da734b14e4CE604f1e18efb7E7f7ef022e96616",
     );
